WebStore/requirements.txt
asgiref                       == 3.8.1
Django                        == 5.1.1
djangorestframework           == 3.15.2
djangorestframework-simplejwt == 5.3.1
django-extensions             == 3.2.3
pillow                        == 10.4.0
pip                           == 24.2
PyJWT                         == 2.9.0
sqlparse                      == 0.5.1
tzdata                        == 2024.1
Werkzeug                      == 3.0.4
pyOpenSSL                     == 24.2.1